src/test/java/uk/gov/hmcts/divorce/caseworker/event/page/AnswerReceivedUploadDocumentTest.java
package uk.gov.hmcts.divorce.caseworker.event.page;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.junit.jupiter.MockitoExtension;
import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
import uk.gov.hmcts.divorce.divorcecase.model.State;
import uk.gov.hmcts.divorce.document.model.DivorceDocument;
import uk.gov.hmcts.divorce.document.model.DocumentType;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.assertNull;
import static uk.gov.hmcts.divorce.testutil.TestDataHelper.LOCAL_DATE_TIME;
import static uk.gov.hmcts.divorce.testutil.TestDataHelper.caseData;

@ExtendWith(MockitoExtension.class)
public class AnswerReceivedUploadDocumentTest {

    @InjectMocks
    private AnswerReceivedUploadDocument page;

    @Test
    public void shouldNotReturnErrorsIfDocumentUploadedIsD11Type() {
        final CaseData caseData = caseData();
        caseData.setD11Document(
            DivorceDocument.builder()
                .documentType(DocumentType.D11)
                .build()
        );

        final CaseDetails<CaseData, State> details = new CaseDetails<>();
        details.setData(caseData);
        details.setId(1L);
        details.setCreatedDate(LOCAL_DATE_TIME);

        AboutToStartOrSubmitResponse<CaseData, State> response = page.midEvent(details, details);

        assertNull(response.getErrors());
    }

    @Test
    public void shouldReturnErrorsIfDocumentUploadedIsNotD11Type() {
        final CaseData caseData = caseData();
        caseData.setD11Document(
            DivorceDocument.builder()
                .documentType(DocumentType.D9H)
                .build()
        );

        final CaseDetails<CaseData, State> details = new CaseDetails<>();
        details.setData(caseData);
        details.setId(1L);
        details.setCreatedDate(LOCAL_DATE_TIME);

        AboutToStartOrSubmitResponse<CaseData, State> response = page.midEvent(details, details);

        assertThat(response.getErrors().size()).isEqualTo(1);
        assertThat(response.getErrors()).contains("Please upload a D11 document type");
    }
}
